requirements.txt
asgiref==3.8.1
black==24.10.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
Django==5.1.3
django-admin-sortable2==2.2.4
django-appconf==1.0.6
django-ckeditor-5==0.2.15
django-ckeditor5==0.0.4
django-debug-toolbar==4.4.6
django-imagekit==5.0.0
django-js-asset==2.2.0
django-solo==2.4.0
djangorestframework==3.15.2
git-filter-repo==2.47.0
gunicorn==23.0.0
idna==3.10
lorem==0.1.1
mypy-extensions==1.0.0
packaging==24.2
pathspec==0.12.1
pilkit==3.0
pillow==11.0.0
platformdirs==4.3.6
psycopg2-binary==2.9.10
python-decouple==3.8
python-http-client==3.3.7
pytz==2024.2
requests==2.32.3
sendgrid==6.11.0
sqlparse==0.5.2
starkbank-ecdsa==2.2.0
stripe==11.3.0
typing_extensions==4.12.2
urllib3==2.2.3
